Make it easy to subclass Saml2Controller
It is normal to want to extend Saml2Controller,
notably to pass your app's own "redirect URL"
to the Saml2Auth login() call, so that the
RelayState will be set accordingly.

This commit makes it easy to do so by adding
an optional config value that lets you specify the controller.
By default, with no value provided, the routes will
be configured as before, with the default Saml2Controller.

Updates the Readme with example of how to do it.