Throw an Error For Invalid IdP Name
Currently when you make a request to `http://myapp.test/saml2/{idpName}/metadata` or any other route with an IdP that you have not registered in the `saml2.php` you will currently get the following error:
```
Undefined index: privateKey
```
The error is caused by the `loadOneLoginAuthFromIpdConfig` attempting to access the private key of a null configuration. This error is ambiguous so to provide a more detailed error we can add a check in the `loadOneLoginAuthFromIdpConfig` which will check to see if that IdP has a config in the `saml2` config folder.